pyproject.toml
[build-system]
requires = ["setuptools>=77.0"]
build-backend = "setuptools.build_meta"

[project]
name = "whirlpool_sixth_sense"
version = "0.20.0"
license = "MIT"
authors = [{ name = "Abílio Costa", email = "amfcalt@gmail.com" }]
description = "Unofficial API for Whirlpool's 6th Sense appliances"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.9.1",
    "websockets>=8.1",
    "async-timeout>=4.0.3",
    "mcp>=1.0.0",
    "uvicorn>=0.24.0",
    "fastapi>=0.104.0",
    "python-dotenv>=1.0.0",
]

[project.readme]
file = "README.md"
content-type = "text/markdown"

[project.urls]
Homepage = "https://github.com/abmantis/whirlpool-sixth-sense/"

[tool.setuptools]
include-package-data = false

[tool.setuptools.packages.find]
exclude = ["tests"]
namespaces = false

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # isort
    "I",]


[tool.ruff.lint.mccabe]
max-complexity = 25

[tool.pyright]
typeCheckingMode = "standard"
